+++ eqcorrscan/utils/src/libutils.h
@@ -58,11 +58,9 @@ int findpeaks(float*, long, float);
 int multi_find_peaks(float*, long, int, float*, int);
 
 // multi_corr functions
-int normxcorr_fftw(float*, long, long, float*, long, float*, long, int*, int*, int*);
-
-
-int normxcorr_fftw_main(float*, long, long, float*, long, float*, long, float*, float*, float*,
-        fftwf_complex*, fftwf_complex*, fftwf_complex*, fftwf_plan, fftwf_plan, fftwf_plan, int*, int*, int, int*);
+int normxcorr_fftw_main(float*, long, long, float*, long, int, int, float*, long, float*, float*, float*,
+                        fftwf_complex*, fftwf_complex*, fftwf_complex*, fftwf_plan, fftwf_plan,
+                        fftwf_plan, int*, int*, int, int*, int);
 
 int normxcorr_fftw_threaded(float*, long, long, float*, long, float*, long, int*, int*, int*);
 
@@ -70,7 +68,10 @@ void free_fftwf_arrays(int, float**, float**, float**, fftwf_complex**, fftwf_co
 
 void free_fftw_arrays(int, double**, double**, double**, fftw_complex**, fftw_complex**, fftw_complex**);
 
-int multi_normxcorr_fftw(float*, long, long, long, float*, long, float*, long, int*, int*, int, int, int*);
+int multi_normxcorr_fftw(float*, long, long, long, float*, long, float*, long, int*, int*, int, int, int*, int);
+
+int normxcorr_fftw(float*, long, long, float*, long, float*, long, int*, int*, int*);
+
 
 // time_corr functions
 int normxcorr_time_threaded(float*, int, float*, int, float*, int);